aider: The function description now specifies that the price should be a number and the location should be a pair of longitude and latitude numbers.
# Aider chat conversation:

USER: the function description is json format, let's leverage that: we want the price to be a number and we want the location to be long, lat numbers as well

ASSISTANT: To make the changes you requested, we need to modify the `app.py` file. Specifically, we need to update the function descriptions for the OpenAI API to specify that the price should be a number and the location should be a pair of numbers representing longitude and latitude.

Here is the *edit block* for the changes:

<source>python
app.py
<<<<<<< HEAD
function_descriptions = [
    {
        "name": "extract_info_from_listing",
        "description": "Categorise & extract key info from a property listing, such as location, price, etc.",
        "parameters": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string",
                    "description": "The location of the property"
                },
                "price": {
                    "type": "string",
                    "description": "The price of the property"
                }
            },
            "required": ["location", "price"]
        }
    }
]
=======
function_descriptions = [
    {
        "name": "extract_info_from_listing",
        "description": "Categorise & extract key info from a property listing, such as location, price, etc.",
        "parameters": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "description": "The longitude and latitude of the property"
                },
                "price": {
                    "type": "number",
                    "description": "The price of the property"
                }
            },
            "required": ["location", "price"]
        }
    }
]
>>>>>>> updated
</source>

This change will ensure that the OpenAI API expects the location to be an array of two numbers (longitude and latitude) and the price to be a number.