aider: Added demo.png screenshot to the `README.md`.
# Aider chat conversation:

USER: please add the demo.png screenshot to the `README.md`

ASSISTANT: To add an image to the `README.md` file, we need to use the Markdown syntax for images. The syntax is `![Alt text](url)`. In this case, the url will be the relative path to the image file in the project directory.

Here is the *edit block* for this change:

<source>markdown
README.md
<<<<<<< HEAD
アプリは `http://localhost:8501` で利用可能です。
=======
アプリは `http://localhost:8501` で利用可能です。

![Demo screenshot](demo.png)
>>>>>>> updated
</source>

This will add a line to the end of the `README.md` file that includes the image. The image will be displayed when the `README.md` is viewed on a platform that supports images in Markdown files, such as GitHub.